Introduce LinkButton to support multiline on iOS 15
